FIX: fix chest render and pause ui
