+++ Assets/Prefabs/Chest.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bfe37316f5eaee34aa0822c141efd4a7
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
