+++ Assets/Scenes/Stage2.unity
@@ -173,6 +173,67 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1001 &51921960
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 4535509029521302781, guid: 90a315c1af6dcff4babcebb268685698, type: 3}
+      propertyPath: m_Camera
+      value: 
+      objectReference: {fileID: 351811198}
+    - target: {fileID: 4535509030702248273, guid: 90a315c1af6dcff4babcebb268685698, type: 3}
+      propertyPath: m_RootOrder
+      value: 9
+      objectReference: {fileID: 0}
+    - target: {fileID: 4535509030702248273, guid: 90a315c1af6dcff4babcebb268685698, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 4.35651
+      objectReference: {fileID: 0}
+    - target: {fileID: 4535509030702248273, guid: 90a315c1af6dcff4babcebb268685698, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0.57343876
+      objectReference: {fileID: 0}
+    - target: {fileID: 4535509030702248273, guid: 90a315c1af6dcff4babcebb268685698, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: -0.17475672
+      objectReference: {fileID: 0}
+    - target: {fileID: 4535509030702248273, guid: 90a315c1af6dcff4babcebb268685698, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 4535509030702248273, guid: 90a315c1af6dcff4babcebb268685698, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4535509030702248273, guid: 90a315c1af6dcff4babcebb268685698, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4535509030702248273, guid: 90a315c1af6dcff4babcebb268685698, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4535509030702248273, guid: 90a315c1af6dcff4babcebb268685698, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4535509030702248273, guid: 90a315c1af6dcff4babcebb268685698, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4535509030702248273, guid: 90a315c1af6dcff4babcebb268685698, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4535509030702248277, guid: 90a315c1af6dcff4babcebb268685698, type: 3}
+      propertyPath: m_Name
+      value: UI
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 90a315c1af6dcff4babcebb268685698, type: 3}
 --- !u!1 &71207834
 GameObject:
   m_ObjectHideFlags: 0
@@ -1704,6 +1765,73 @@ Tilemap:
     e31: 0
     e32: 0
     e33: 1
+--- !u!1 &528101414
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 528101417}
+  - component: {fileID: 528101416}
+  - component: {fileID: 528101415}
+  m_Layer: 0
+  m_Name: EventSystem
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &528101415
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 528101414}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4f231c4fb786f3946a6b90b886c48677, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_HorizontalAxis: Horizontal
+  m_VerticalAxis: Vertical
+  m_SubmitButton: Submit
+  m_CancelButton: Cancel
+  m_InputActionsPerSecond: 10
+  m_RepeatDelay: 0.5
+  m_ForceModuleActive: 0
+--- !u!114 &528101416
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 528101414}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 76c392e42b5098c458856cdf6ecaaaa1, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_FirstSelected: {fileID: 0}
+  m_sendNavigationEvents: 1
+  m_DragThreshold: 10
+--- !u!4 &528101417
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 528101414}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 10
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &685365747
 GameObject:
   m_ObjectHideFlags: 0
@@ -2071,6 +2199,63 @@ Transform:
   m_Father: {fileID: 1611684205}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1001 &1032669139
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 3352970000075848385, guid: bfe37316f5eaee34aa0822c141efd4a7, type: 3}
+      propertyPath: m_Name
+      value: Chest
+      objectReference: {fileID: 0}
+    - target: {fileID: 3352970000075848398, guid: bfe37316f5eaee34aa0822c141efd4a7, type: 3}
+      propertyPath: m_RootOrder
+      value: 8
+      objectReference: {fileID: 0}
+    - target: {fileID: 3352970000075848398, guid: bfe37316f5eaee34aa0822c141efd4a7, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: -19
+      objectReference: {fileID: 0}
+    - target: {fileID: 3352970000075848398, guid: bfe37316f5eaee34aa0822c141efd4a7, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 19.5
+      objectReference: {fileID: 0}
+    - target: {fileID: 3352970000075848398, guid: bfe37316f5eaee34aa0822c141efd4a7, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3352970000075848398, guid: bfe37316f5eaee34aa0822c141efd4a7, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3352970000075848398, guid: bfe37316f5eaee34aa0822c141efd4a7, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3352970000075848398, guid: bfe37316f5eaee34aa0822c141efd4a7, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3352970000075848398, guid: bfe37316f5eaee34aa0822c141efd4a7, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3352970000075848398, guid: bfe37316f5eaee34aa0822c141efd4a7, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3352970000075848398, guid: bfe37316f5eaee34aa0822c141efd4a7, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3352970000075848398, guid: bfe37316f5eaee34aa0822c141efd4a7, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: bfe37316f5eaee34aa0822c141efd4a7, type: 3}
 --- !u!1 &1069836963
 GameObject:
   m_ObjectHideFlags: 0
