+++ Assets/Scripts/Character/PlayerController.cs
@@ -9,7 +9,6 @@ namespace Character
     public class PlayerController : MonoBehaviour
     {
         private CharacterController2D _mCharacterController;
-        public static bool GameStop = false;
         private void Awake()
         {
             _mCharacterController = GetComponent<CharacterController2D>();
@@ -43,20 +42,6 @@ namespace Character
             if (Input.GetKeyDown(KeyCode.Q))
             {
                 _mCharacterController.TakeDamage(gameObject, 1);
-            }
-            if (Input.GetKeyDown(KeyCode.P))
-            {
-                ///Debug.Log("stopGame");
-                GameStop = !GameStop;
-                if (GameStop)
-                {    
-                    Time.timeScale = 0;
-                }
-                else
-                {
-                    Time.timeScale = 1;
-                }
-            
             }
             
         }
