+++ Assets/Scripts/UI/BeginUIChange.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 21ba37db5c0beac4ebfba1ff15fb852a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
