+++ Assets/Scripts/UI/StopMenuShow.cs
@@ -1,12 +1,12 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using Character;
+using UnityEngine.Events;
 
 public class StopMenuShow : MonoBehaviour
 {
     // Start is called before the first frame update
-    private GameObject[] UI;
+    [SerializeField] [Tooltip("交互事件列表。")] private UnityEvent keyDownEvent = null;
     void Start()
     {
         
@@ -15,17 +15,9 @@ public class StopMenuShow : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        if(PlayerController.GameStop)
+        if (Input.GetKeyUp(KeyCode.P))
         {
-            transform.Find("quit").gameObject.SetActive(true);
-            transform.Find("startBtn").gameObject.SetActive(true);
-            transform.Find("restart").gameObject.SetActive(true);
-        }
-        else
-        {
-            transform.Find("quit").gameObject.SetActive(false);
-            transform.Find("startBtn").gameObject.SetActive(false);
-            transform.Find("restart").gameObject.SetActive(false);
+            keyDownEvent.Invoke();
         }
         
     }
