+++ ProjectSettings/EditorBuildSettings.asset
@@ -6,9 +6,6 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
-    path: Assets/Scenes/SampleScene.unity
-    guid: 2cda990e2423bbf4892e6590ba056729
-  - enabled: 1
     path: Assets/Scenes/BeginScene.unity
     guid: 097e1f36312a0f142bde69245d0cc1ee
   - enabled: 1
@@ -17,4 +14,7 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/Stage1.unity
     guid: 735481a80101fbd42b705665b6cd586f
+  - enabled: 1
+    path: Assets/Scenes/Stage2.unity
+    guid: 32fb32e148680e446a362b79556572ba
   m_configObjects: {}
